pyproject.toml
[project]
name = "cdk-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.189.0",
    "constructs>=10.0.0,<11.0.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.399",
    "pytest>=8.3.5",
    "ruff>=0.11.5",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
show-fixes = true
line-length = 100

[tool.ruff.lint]
extend-select = ["I", "RUF", "T20"]
ignore = ["RUF012"]

